Derrive counters as double
This fixes counters always being 0 when you have many threads. For
instance, 10 connects per second on 8 threads is 0.8 per second per
thread, which was counted as 0. And 0+0+0... = 0.

+++ derivablecounter.cpp
@@ -27,14 +27,14 @@ uint64_t DerivableCounter::get() const
  * Even though the class it not meant to be thread-safe, this method does use a mutex, because obtaining the value can be
  * scheduled in different threads.
  */
-uint64_t DerivableCounter::getPerSecond()
+double DerivableCounter::getPerSecond()
 {
     std::lock_guard<std::mutex> locker(timeMutex);
 
     std::chrono::time_point<std::chrono::steady_clock> now = std::chrono::steady_clock::now();
     std::chrono::milliseconds msSinceLastTime = std::chrono::duration_cast<std::chrono::milliseconds>(now - timeOfPrevious);
     uint64_t messagesTimes1000 = (val - valPrevious) * 1000;
-    uint64_t result = messagesTimes1000 / (msSinceLastTime.count() + 1); // branchless avoidance of div by 0;
+    double result = messagesTimes1000 / static_cast<double>(msSinceLastTime.count() + 1); // branchless avoidance of div by 0;
     timeOfPrevious = now;
     valPrevious = val;
     return result;
